requirements.txt
ansiwrap==0.8.4
appdirs @ file:///home/conda/feedstock_root/build_artifacts/appdirs_1603108395799/work
audioread @ file:///home/conda/feedstock_root/build_artifacts/audioread_1667036998326/work
biopython @ file:///home/conda/feedstock_root/build_artifacts/biopython_1666879192007/work
brotlipy @ file:///home/conda/feedstock_root/build_artifacts/brotlipy_1666764671472/work
cairocffi==1.4.0
certifi==2022.12.7
cffi @ file:///home/conda/feedstock_root/build_artifacts/cffi_1671179353105/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1661170624537/work
click @ file:///home/conda/feedstock_root/build_artifacts/click_1666798198223/work
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1666700638685/work
contourpy @ file:///home/conda/feedstock_root/build_artifacts/contourpy_1673633665736/work
cryptography @ file:///home/conda/feedstock_root/build_artifacts/cryptography-split_1675828607645/work
cssselect2==0.7.0
cycler @ file:///home/conda/feedstock_root/build_artifacts/cycler_1635519461629/work
Cython @ file:///home/conda/feedstock_root/build_artifacts/cython_1673054058583/work
dataclassy @ file:///home/conda/feedstock_root/build_artifacts/dataclassy_1673711725182/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
defusedxml==0.7.1
fonttools @ file:///home/conda/feedstock_root/build_artifacts/fonttools_1666827107856/work
greenlet @ file:///home/conda/feedstock_root/build_artifacts/greenlet_1674937639600/work
hdbscan @ file:///home/conda/feedstock_root/build_artifacts/hdbscan_1675623815609/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1663625384323/work
joblib @ file:///home/conda/feedstock_root/build_artifacts/joblib_1663332044897/work
kiwisolver @ file:///home/conda/feedstock_root/build_artifacts/kiwisolver_1666805701884/work
kneed @ file:///home/conda/feedstock_root/build_artifacts/kneed_1597304056973/work
librosa @ file:///home/conda/feedstock_root/build_artifacts/librosa_1656336149289/work
llvmlite==0.39.1
lxml==4.9.2
matplotlib @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-suite_1674078815158/work
Montreal-Forced-Aligner @ file:///home/conda/feedstock_root/build_artifacts/montreal-forced-aligner_1675734778408/work
munkres==1.1.4
numba @ file:///home/conda/feedstock_root/build_artifacts/numba_1669103970038/work
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1668919096861/work
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1673482170163/work
pandas==1.5.3
patsy @ file:///home/conda/feedstock_root/build_artifacts/patsy_1665356157073/work
pgvector @ file:///home/conda/feedstock_root/build_artifacts/pgvector-python_1675578182088/work
Pillow @ file:///home/conda/feedstock_root/build_artifacts/pillow_1675487172403/work
pooch @ file:///home/conda/feedstock_root/build_artifacts/pooch_1643032624649/work
praat-parselmouth==0.4.3
praatio @ file:///home/conda/feedstock_root/build_artifacts/praatio_1675613839329/work
psycopg2 @ file:///home/conda/feedstock_root/build_artifacts/psycopg2-split_1672159192599/work
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1636257122734/work
pynini @ file:///home/conda/feedstock_root/build_artifacts/pynini_1667113453535/work
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1672659226110/work
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/pyparsing_1652235407899/work
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1661604839144/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1673864280276/work
PyYAML @ file:///home/conda/feedstock_root/build_artifacts/pyyaml_1666772395347/work
reportlab==3.6.12
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1673863902341/work
resampy @ file:///home/conda/feedstock_root/build_artifacts/resampy_1663087224680/work
scikit-learn @ file:///home/conda/feedstock_root/build_artifacts/scikit-learn_1674598990422/work
scipy==1.10.0
seaborn @ file:///home/conda/feedstock_root/build_artifacts/seaborn-split_1672497695270/work
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
soundfile @ file:///home/conda/feedstock_root/build_artifacts/pysoundfile_1665058838115/work
SQLAlchemy @ file:///home/conda/feedstock_root/build_artifacts/sqlalchemy_1675747090683/work
statsmodels @ file:///home/conda/feedstock_root/build_artifacts/statsmodels_1667586015456/work
textwrap3==0.9.2
threadpoolctl @ file:///home/conda/feedstock_root/build_artifacts/threadpoolctl_1643647933166/work
tinycss2==1.2.1
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1662214488106/work
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1665144421445/work
unicodedata2 @ file:///home/conda/feedstock_root/build_artifacts/unicodedata2_1667239886688/work
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1673452138552/work
webencodings==0.5.1
